refactor(redesign): refactor multiple quote fetcher hook
